Add throttling to DispatchQueue and improve debugability with SourceInfo Clean up async renderable loading Improve read-only editor serialization to be copyable Fix DispatchQueue recursive bug

+++ src/core/ecs/components/Renderable.cc
@@ -9,6 +9,7 @@
 
 #include "assets/AssetManager.hh"
 
+#include <cstdint>
 #include <picojson.h>
 
 namespace ecs {
@@ -20,7 +21,7 @@ namespace ecs {
         if (dst.joints.empty()) dst.joints = src.joints;
     }
 
-    Renderable::Renderable(std::string_view modelName, uint64_t meshIndex)
+    Renderable::Renderable(std::string_view modelName, uint32_t meshIndex)
         : modelName(modelName), meshIndex(meshIndex) {
         if (modelName.empty()) {
             visibility = VisibilityMask::None;
